# contracts/cloud/openapi.yaml
openapi: 3.1.0
info:
  title: AO Cloud Public API
  version: 1.0.0
  description: |
    Client-facing contract for a separate AO Cloud implementation. This
    repository does not serve these routes. The worker surface documents the
    exact outbound HTTP protocol used by sandbox workers, including one-time
    bootstrap, rotating credentials, fenced turn execution, orchestration, and
    durable workspace and terminal transport.
security:
  - bearerAuth: []
...
paths:
...
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/messages:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    post:
      operationId: sendMessage
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSendMessageInput"
      responses:
        "202":
          description: Message accepted and durably appended.
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event:
                    $ref: "#/components/schemas/UserMessageEvent"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/turns/{turnId}/cancel:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/TurnId"
    post:
      operationId: cancelTurn
      tags: [Sessions]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Cancellation was durably requested for the turn.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerOKResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/chat-events:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: replayClientEvents
      tags: [Events]
      parameters:
        - $ref: "#/components/parameters/After"
        - $ref: "#/components/parameters/EventLimit"
      responses:
        "200":
          description: Client-visible events after the supplied sequence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientEventPage"
        default:
          $ref: "#/components/responses/Error"
  /api/cloud/v1/orgs/{orgId}/sessions/{sessionId}/events:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/SessionId"
    get:
      operationId: streamClientEvents
      tags: [Events]
      parameters:
        - $ref: "#/components/parameters/After"
      responses:
        "200":
          description: |
            Replay followed by live client-visible events. Each SSE data field
            contains a ClientEvent; the SSE id is its sequence.
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    workerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Session-scoped credential formatted as `Worker <token>`. Every
        authenticated route verifies both the token signature and that its
        worker ID and epoch are still current. Replacing a sandbox therefore
        fences an otherwise unexpired token with `STALE_WORKER_TOKEN`.

        Route scopes are `worker:connect` for heartbeat, `worker:event` for
        events, `worker:turn:claim`, `worker:turn:poll`, and
        `worker:turn:complete` for turn execution,
        `worker:credential:read` for the agent secret, `worker:git` for a
        checkout grant, `worker:orchestrate` for child controls, and
        `worker:transport` for workspace and terminal transport.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TurnId:
      name: turnId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    EventLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    After:
      name: after
      in: query
      description: Return events with a sequence strictly greater than this value.
      schema:
        type: integer
        format: int64
        minimum: 0
        maximum: 9007199254740991
        default: 0
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Reusing a key with the same command returns the original result.
        Reusing it with a different command returns an IDEMPOTENCY_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
  responses:
    Error:
      description: Standard Cloud API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, code, message, requestId]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
...
    WorkerOutputPayload:
      type: object
      additionalProperties: false
      required: [turnId, attempt, stream, text]
      properties:
        turnId:
          type: string
          format: uuid
        attempt:
          type: integer
          minimum: 1
        stream:
          type: string
          enum: [stdout, stderr]
        text:
          type: string
          minLength: 1
          description: Assistant output frame, limited by the server to 16 KiB.
...
    WorkerOKResponse:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
...
    UserSendMessageInput:
      type: object
      additionalProperties: false
      required: [text, clientSequence]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 65536
        clientSequence:
          type: integer
          minimum: 1
          maximum: 9007199254740991
...
    UserMessageEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.user_message
        payload:
          type: object
          additionalProperties: false
          required: [text]
          properties:
            text:
              type: string
        createdAt:
          type: string
          format: date-time
    AssistantDeltaEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.assistant_delta
        payload:
          $ref: "#/components/schemas/WorkerOutputPayload"
        createdAt:
          type: string
          format: date-time
    TurnStartedEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.turn_started
        payload:
          type: object
          additionalProperties: false
          required: [turnId, attempt, workerEpoch]
          properties:
            turnId:
              type: string
              format: uuid
            attempt:
              type: integer
              minimum: 1
            workerEpoch:
              type: integer
              format: int64
              minimum: 1
        createdAt:
          type: string
          format: date-time
    TurnCompletedEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.turn_completed
        payload:
          type: object
          additionalProperties: false
          required: [turnId, attempt]
          properties:
            turnId:
              type: string
              format: uuid
            attempt:
              type: integer
              minimum: 1
        createdAt:
          type: string
          format: date-time
    TurnInterruptedEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.turn_interrupted
        payload:
          type: object
          additionalProperties: false
          required: [turnId, attempt]
          properties:
            turnId:
              type: string
              format: uuid
            attempt:
              type: integer
              minimum: 1
        createdAt:
          type: string
          format: date-time
    TurnAbortedEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.turn_aborted
        payload:
          type: object
          additionalProperties: false
          required: [turnId, attempt, error]
          properties:
            turnId:
              type: string
              format: uuid
            attempt:
              type: integer
              minimum: 1
            error:
              type: string
        createdAt:
          type: string
          format: date-time
    InterruptRequestedEvent:
      type: object
      additionalProperties: false
      required: [sessionId, sequence, type, payload, createdAt]
      properties:
        sessionId:
          type: string
        sequence:
          type: integer
          format: int64
          maximum: 9007199254740991
        type:
          const: chat.interrupt_requested
        payload:
          type: object
          additionalProperties: false
          required: [turnId]
          properties:
            turnId:
              type: string
              format: uuid
        createdAt:
          type: string
          format: date-time
    ClientEvent:
      oneOf:
        - $ref: "#/components/schemas/UserMessageEvent"
        - $ref: "#/components/schemas/AssistantDeltaEvent"
        - $ref: "#/components/schemas/TurnStartedEvent"
        - $ref: "#/components/schemas/TurnCompletedEvent"
        - $ref: "#/components/schemas/TurnInterruptedEvent"
        - $ref: "#/components/schemas/TurnAbortedEvent"
        - $ref: "#/components/schemas/InterruptRequestedEvent"
    ClientEventPage:
      type: object
      additionalProperties: false
      required: [events, hasMore, nextAfter]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ClientEvent"
        hasMore:
          type: boolean
        nextAfter:
          type: integer
          format: int64
          minimum: 0
          maximum: 9007199254740991
          description: Highest sequence returned, or the supplied after cursor when the page is empty.
...